backend/requirements.txt
fastapi
uvicorn
pydantic
PyGithub
openai
python-multipart
markdown
